+++ src/components/_inputs/InputBoolean.vue
@@ -1,10 +1,20 @@
 <template>
-  <a-switch
-    :name="name"
-    :class="localValue && name.indexOf('disable') > -1 ? 'bg-warning' : null"
-    :defaultChecked="Boolean(schema.default)"
-    v-model="localValue"
-  />
+  <div class="custom-control custom-checkbox">
+    <input
+      v-model="localValue"
+      type="checkbox"
+      class="custom-control-input"
+      :id="schema.title"
+      :class="localValue && isDisable ? 'bg-warning' : null"
+    >
+    <label
+      class="custom-control-label"
+      :class="isDisable ? 'text-warning' : null"
+      :for="schema.title"
+    >
+      {{ schema.title }}
+    </label>
+  </div>
 </template>
 
 <script src="./js/InputBoolean.js"></script>