+++ src/components/_inputs/InputEnum.vue
@@ -1,17 +1,17 @@
 <template>
-  <a-select
-    style="min-width: 120px"
-    :default-value="schema.default"
+  <select
+    class="custom-select"
+    style="min-width: 130px"
     v-model="localValue"
   >
-    <a-select-option
+    <option
       v-for="item in schema.enum"
       :key="item"
       :value="item"
     >
       {{ i18nValues ? i18nValues[item] : item}}
-    </a-select-option>
-  </a-select>
+    </option>
+  </select>
 </template>
 
 <script src="./js/InputEnum.js"></script>