+++ src/components/_inputs/InputList.vue
@@ -1,6 +1,7 @@
 <template>
   <div>
     <input-text
+      class="mb-2"
       v-for="(item, i) in localValue"
       :name="`${name}[${i}]`"
       :key="i"
@@ -9,15 +10,13 @@
       @input="value => setItemValue(value, i)"
       @blur="checkItem(i)"
     />
-    <a-button
-      type="dashed"
-      html-type="button"
-      size="small"
+    <button
+      class="mb-1 btn btn-secundary btn-sm"
       @click.prevent="addItem()"
     >
       <i class="fas fa-plus-circle mr-1"></i>
       {{ schema.items.title }}
-    </a-button>
+    </button>
   </div>
 </template>