+++ src/components/_inputs/js/InputNumber.js
@@ -1,5 +1,3 @@
-import { $ecomConfig } from '@ecomplus/utils'
-
 export default {
   name: 'InputNumber',
 
@@ -18,8 +16,6 @@ export default {
   },
 
   computed: {
-    countryCode: () => $ecomConfig.get('country_code'),
-
     localValue: {
       get () {
         const num = parseFloat(this.value)
@@ -28,6 +24,11 @@ export default {
       set (val) {
         this.$emit('input', val)
       }
+    },
+
+    step () {
+      return this.schema.type === 'integer' ? 1
+        : (this.schema.multipleOf || 0.01)
     }
   },