+++ src/components/_inputs/js/InputZipCode.js
@@ -17,7 +17,9 @@ export default {
     },
     schema: {
       type: Object,
-      default: () => ({})
+      default () {
+        return {}
+      }
     }
   },