+++ src/lib/_custom-input-matchers/zip-range.js
@@ -1,0 +1,13 @@
+import InputZipRange from './../../components/_inputs/InputZipRange.vue'
+
+export default (field, localSchema) => {
+  if (field.includes('zip')) {
+    const { properties } = localSchema
+    if (properties && properties.min && properties.max) {
+      delete properties.min
+      delete properties.max
+      return InputZipRange
+    }
+  }
+  return false
+}